refactor: OrderTable 패키지 분리
