+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -20,6 +20,8 @@ import javax.persistence.Table;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import kitchenpos.table.domain.OrderTable;
+
 @Entity
 @EntityListeners(AuditingEntityListener.class)
 @Table(name = "orders")
